refactor(25464): refactor the toolbar button group for reuse
