+++ hivemq-edge/src/frontend/src/extensions/datahub/components/nodes/NodeDatahubToolbar.tsx
@@ -1,10 +1,10 @@
-import { FC, useMemo } from 'react'
+import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
-import { ButtonGroup, ButtonGroupProps } from '@chakra-ui/react'
+import { ButtonGroupProps } from '@chakra-ui/react'
 import { LuCopy, LuDelete, LuFileEdit } from 'react-icons/lu'
 
 import IconButton from '@/components/Chakra/IconButton.tsx'
-import { useStore } from 'reactflow'
+import ToolbarButtonGroup from '@/components/react-flow/ToolbarButtonGroup.tsx'
 
 interface NodeToolbarProps extends ButtonGroupProps {
   onCopy?: (event: React.BaseSyntheticEvent) => void
@@ -14,17 +14,9 @@ interface NodeToolbarProps extends ButtonGroupProps {
 
 const NodeDatahubToolbar: FC<NodeToolbarProps> = (props) => {
   const { t } = useTranslation('datahub')
-  const zoomFactor = useStore((s) => s.transform[2])
-
-  const getToolbarSize = useMemo<string>(() => {
-    if (zoomFactor >= 1.5) return 'lg'
-    if (zoomFactor >= 1) return 'md'
-    if (zoomFactor >= 0.75) return 'sm'
-    return 'xs'
-  }, [zoomFactor])
 
   return (
-    <ButtonGroup size={getToolbarSize} variant="solid" colorScheme="gray" isAttached {...props}>
+    <ToolbarButtonGroup orientation="horizontal" isAttached {...props}>
       <IconButton
         icon={<LuFileEdit />}
         data-testid="node-toolbar-edit"
@@ -44,7 +36,7 @@ const NodeDatahubToolbar: FC<NodeToolbarProps> = (props) => {
         colorScheme="red"
         onClick={props.onDelete}
       />
-    </ButtonGroup>
+    </ToolbarButtonGroup>
   )
 }
