+++ hivemq-edge/src/frontend/src/modules/Workspace/components/nodes/NodeAdapter.tsx
@@ -8,6 +8,7 @@ import { type Adapter } from '@/api/__generated__'
 import { useGetAdapterTypes } from '@/api/hooks/useProtocolAdapters/useGetAdapterTypes.ts'
 import IconButton from '@/components/Chakra/IconButton.tsx'
 import { ConnectionStatusBadge } from '@/components/ConnectionStatusBadge'
+import ToolbarButtonGroup from '@/components/react-flow/ToolbarButtonGroup.tsx'
 import { type TopicFilter } from '@/modules/Workspace/types.ts'
 import { useEdgeFlowContext } from '@/modules/Workspace/hooks/useEdgeFlowContext.ts'
 import { discoverAdapterTopics } from '@/modules/Workspace/utils/topics-utils.ts'
@@ -17,7 +18,6 @@ import ContextualToolbar from '@/modules/Workspace/components/nodes/ContextualTo
 import NodeWrapper from '@/modules/Workspace/components/parts/NodeWrapper.tsx'
 import TopicsContainer from '@/modules/Workspace/components/parts/TopicsContainer.tsx'
 import { CONFIG_ADAPTER_WIDTH } from '@/modules/Workspace/utils/nodes-utils.ts'
-import WorkspaceButtonGroup from '@/modules/Workspace/components/parts/WorkspaceButtonGroup.tsx'
 import { selectorIsSkeletonZoom } from '@/modules/Workspace/utils/react-flow.utils.ts'
 
 const NodeAdapter: FC<NodeProps<Adapter>> = ({ id, data: adapter, selected, dragging }) => {
@@ -40,7 +40,7 @@ const NodeAdapter: FC<NodeProps<Adapter>> = ({ id, data: adapter, selected, drag
   return (
     <>
       <ContextualToolbar id={id} dragging={dragging} onOpenPanel={onContextMenu}>
-        <WorkspaceButtonGroup>
+        <ToolbarButtonGroup>
           {HACK_BIDIRECTIONAL && (
             <IconButton
               icon={<Icon as={deviceCapabilityIcon['WRITE']} />}
@@ -53,7 +53,7 @@ const NodeAdapter: FC<NodeProps<Adapter>> = ({ id, data: adapter, selected, drag
             aria-label={t('workspace.toolbar.command.subscriptions.inward')}
             onClick={() => navigate(`${adapterNavPath}/inward`)}
           />
-        </WorkspaceButtonGroup>
+        </ToolbarButtonGroup>
       </ContextualToolbar>
       <NodeWrapper
         isSelected={selected}
